refactor: unify DDL execution and dry-run into shared processCreate* methods
Extract TemporaryTable to a standalone public class. Add DdlMode enum
(CREATE/UPDATE/SPECIFY), processCreateMaterializedView(), and
processCreateTable() to HoptimatorDdlUtils so HoptimatorDdlExecutor
fully delegates to them. Add specifyFromSql() and specifyCreateTable()
as the unified !specify entry points used by QuidemTestBase and
HoptimatorAppConfig.
