+++ hoptimator-jdbc/src/main/java/com/linkedin/hoptimator/jdbc/HoptimatorDdlExecutor.java
@@ -19,11 +19,7 @@
  */
 package com.linkedin.hoptimator.jdbc;
 
-import com.google.common.collect.ImmutableList;
-import com.linkedin.hoptimator.Database;
 import com.linkedin.hoptimator.Deployer;
-import com.linkedin.hoptimator.MaterializedView;
-import com.linkedin.hoptimator.Pipeline;
 import com.linkedin.hoptimator.Source;
 import com.linkedin.hoptimator.Trigger;
 import com.linkedin.hoptimator.UserJob;
@@ -35,22 +31,11 @@ import com.linkedin.hoptimator.jdbc.ddl.SqlCreateTrigger;
 import com.linkedin.hoptimator.jdbc.ddl.SqlDropTrigger;
 import com.linkedin.hoptimator.jdbc.ddl.SqlPauseTrigger;
 import com.linkedin.hoptimator.jdbc.ddl.SqlResumeTrigger;
-import com.linkedin.hoptimator.util.ArrayTable;
 import com.linkedin.hoptimator.util.DeploymentService;
-import com.linkedin.hoptimator.util.planner.HoptimatorJdbcCatalogSchema;
 import com.linkedin.hoptimator.util.planner.HoptimatorJdbcSchema;
 import com.linkedin.hoptimator.util.planner.HoptimatorJdbcTable;
-import com.linkedin.hoptimator.util.planner.PipelineRel;
-import org.apache.calcite.adapter.java.JavaTypeFactory;
 import org.apache.calcite.jdbc.CalcitePrepare;
 import org.apache.calcite.jdbc.CalciteSchema;
-import org.apache.calcite.jdbc.ContextSqlValidator;
-import org.apache.calcite.plan.RelOptTable;
-import org.apache.calcite.rel.RelRoot;
-import org.apache.calcite.rel.type.RelDataType;
-import org.apache.calcite.rel.type.RelDataTypeFactory;
-import org.apache.calcite.rex.RexNode;
-import org.apache.calcite.schema.ColumnStrategy;
 import org.apache.calcite.schema.Function;
 import org.apache.calcite.schema.SchemaPlus;
 import org.apache.calcite.schema.Table;
@@ -61,25 +46,18 @@ import org.apache.calcite.sql.SqlIdentifier;
 import org.apache.calcite.sql.SqlKind;
 import org.apache.calcite.sql.SqlLiteral;
 import org.apache.calcite.sql.SqlNode;
-import org.apache.calcite.sql.ddl.SqlColumnDeclaration;
 import org.apache.calcite.sql.ddl.SqlCreateView;
 import org.apache.calcite.sql.ddl.SqlDropMaterializedView;
 import org.apache.calcite.sql.ddl.SqlDropObject;
 import org.apache.calcite.sql.ddl.SqlDropTable;
 import org.apache.calcite.sql.ddl.SqlDropView;
-import org.apache.calcite.sql.ddl.SqlKeyConstraint;
 import org.apache.calcite.sql.dialect.CalciteSqlDialect;
 import org.apache.calcite.sql.parser.SqlAbstractParserImpl;
 import org.apache.calcite.sql.parser.SqlParserImplFactory;
 import org.apache.calcite.sql.parser.SqlParserPos;
 import org.apache.calcite.sql.validate.SqlConformanceEnum;
-import org.apache.calcite.sql.validate.SqlValidator;
-import org.apache.calcite.sql2rel.InitializerContext;
-import org.apache.calcite.sql2rel.InitializerExpressionFactory;
-import org.apache.calcite.sql2rel.NullInitializerExpressionFactory;
 import org.apache.calcite.util.Pair;
 
-import javax.annotation.Nullable;
 import java.io.Reader;
 import java.sql.SQLException;
 import java.util.ArrayList;
@@ -88,10 +66,6 @@ import java.util.Collections;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-import java.util.Properties;
-
-import static com.google.common.base.Preconditions.checkArgument;
-import static java.util.Objects.requireNonNull;
 
 
 public final class HoptimatorDdlExecutor extends ServerDdlExecutor {
@@ -200,108 +174,20 @@ public final class HoptimatorDdlExecutor extends ServerDdlExecutor {
   /** Executes a {@code CREATE MATERIALIZED VIEW} command. */
   public void execute(SqlCreateMaterializedView create, CalcitePrepare.Context context) {
     logger.info("Validating statement: {}", create);
+    HoptimatorDdlUtils.DdlMode mode = create.getReplace()
+        ? HoptimatorDdlUtils.DdlMode.UPDATE : HoptimatorDdlUtils.DdlMode.CREATE;
     try {
-      ValidationService.validateOrThrow(create);
-    } catch (SQLException e) {
-      throw new DdlException(create, e.getMessage(), e);
-    }
-
-    final Pair<CalciteSchema, String> pair = HoptimatorDdlUtils.schema(context, true, create.name);
-    if (pair.left == null) {
-      throw new DdlException(create, "Schema for " + create.name + " not found.");
-    }
-    final SchemaPlus schemaPlus = pair.left.plus();
-    if (schemaPlus.tables().get(pair.right) != null) {
-      if (schemaPlus.tables().get(pair.right) instanceof HoptimatorJdbcTable) {
-        throw new DdlException(create,
-            "Cannot overwrite physical table " + pair.right + " with a view.");
-      }
-      // Materialized view exists.
-      if (!create.ifNotExists && !create.getReplace()) {
-        // They did not specify IF NOT EXISTS, so give error.
-        throw new DdlException(create,
-            "View " + pair.right + " already exists. Use CREATE OR REPLACE to update.");
-      }
-      if (create.getReplace()) {
-        schemaPlus.removeTable(pair.right);
-      } else {
-        // nothing to do
-        return;
-      }
-    }
-
-    final SqlNode q = HoptimatorDdlUtils.renameColumns(create.columnList, create.query);
-    final String sql = q.toSqlString(CalciteSqlDialect.DEFAULT).getSql();
-    final List<String> schemaPath = pair.left.path(null);
-
-    Collection<Deployer> deployers = null;
-
-    String schemaName = schemaPlus.getName();
-    String viewName = pair.right;
-    List<String> viewPath = new ArrayList<>(schemaPath);
-    viewPath.add(viewName);
-
-    Pair<SchemaPlus, Table> schemaSnapshot = null;
-    if (!(pair.left.schema instanceof Database)) {
-      throw new DdlException(create, schemaName + " is not a physical database.");
-    }
-    String database = ((Database) pair.left.schema).databaseName();
-    try {
-      logger.info("Validated sql statement. The view is named {} and has path {}",
-          viewName, viewPath);
-
-      // Support "partial views", i.e. CREATE VIEW FOO$BAR, where the view name
-      // is "foo-bar" and the sink is just FOO.
-      String[] viewParts = viewName.split("\\$", 2);
-      String sinkName = viewParts[0];
-      String pipelineName = database + "-" + sinkName;
-      if (viewParts.length > 1) {
-        pipelineName = pipelineName + "-" + viewParts[1];
-      }
-      logger.info("Pipeline name for view {} is {}", viewName, pipelineName);
-      Properties connectionProperties = connection.connectionProperties();
-      connectionProperties.setProperty(DeploymentService.PIPELINE_OPTION, pipelineName);
-
-      // Plan a pipeline to materialize the view.
-      RelRoot root = new HoptimatorDriver.Prepare(connection).convert(context, sql).root;
-      PipelineRel.Implementor plan = DeploymentService.plan(root, connection.materializations(), connectionProperties);
-      schemaSnapshot = HoptimatorDdlUtils.snapshotAndSetSinkSchema(context, new HoptimatorDriver.Prepare(connection), plan, sql, pair);
-      logger.info("Added materialized view {} to schema {}", viewName, schemaPlus.getName());
-      Pipeline pipeline = plan.pipeline(viewName, connection);
-      MaterializedView hook = new MaterializedView(database, viewPath, sql, pipeline.job().sql(), pipeline);
-      // TODO support CREATE ... WITH (options...)
-      logger.info("Validating materialized view {}", viewName);
-      ValidationService.validateOrThrow(hook);
-      deployers = DeploymentService.deployers(hook, connection);
-      logger.info("Validating deployable resources for materialized view {}", viewName);
-      ValidationService.validateOrThrow(deployers);
-      logger.info("Validated materialized view {}", viewName);
-      if (create.getReplace()) {
-        logger.info("Deploying update materialized view {}", viewName);
-        DeploymentService.update(deployers);
-      } else {
-        logger.info("Deploying create materialized view {}", viewName);
-        DeploymentService.create(deployers);
-      }
-      logger.info("Deployed materialized view {}", viewName);
+      HoptimatorDdlUtils.processCreateMaterializedView(
+          context,
+          new HoptimatorDriver.Prepare(connection),
+          connection,
+          create,
+          mode);
     } catch (SQLException | RuntimeException e) {
-      logger.info("Failed to deploy materialized view {}", viewName);
-      if (deployers != null) {
-        DeploymentService.restore(deployers);
-        logger.info("Restored deployable resources for materialized view {}", viewName);
-      }
-      if (schemaSnapshot != null) {
-        if (schemaSnapshot.right == null) {
-          schemaSnapshot.left.removeTable(viewName);
-          logger.info("Removed schema for materialized view {}", viewName);
-        } else {
-          schemaPlus.add(viewName, schemaSnapshot.right);
-          logger.info("Restored schema for materialized view {}", viewName);
-        }
-      }
+      logger.info("Failed to deploy materialized view {}", create.name);
       throw new DdlException(create, e.getMessage(), e);
     }
-    logger.info("CREATE MATERIALIZED VIEW {} completed", viewName);
+    logger.info("CREATE MATERIALIZED VIEW {} completed", create.name);
   }
 
   /** Executes a {@code CREATE TRIGGER} command. */
@@ -371,174 +257,15 @@ public final class HoptimatorDdlExecutor extends ServerDdlExecutor {
 
   /** Executes a {@code CREATE TABLE} command. */
   public void execute(SqlCreateTable create, CalcitePrepare.Context context) {
-    logger.info("Validating statement: {}", create);
+    HoptimatorDdlUtils.DdlMode mode = create.getReplace()
+        ? HoptimatorDdlUtils.DdlMode.UPDATE : HoptimatorDdlUtils.DdlMode.CREATE;
     try {
-      ValidationService.validateOrThrow(create);
-    } catch (SQLException e) {
-      throw new DdlException(create, e.getMessage(), e);
-    }
-
-    boolean isNewSchema = false;
-    Pair<CalciteSchema, String> pair = HoptimatorDdlUtils.schema(context, true, create.name);
-    if (pair.left == null) {
-      // If the schema is not found, it might be because it's a 3-level path (CATALOG.SCHEMA.TABLE)
-      if (create.name.names.size() > 2) {
-        pair = HoptimatorDdlUtils.catalog(context, true, create.name);
-        isNewSchema = true;
-        if (pair.left == null) {
-          throw new DdlException(create, "Catalog for " + create.name + " not found.");
-        }
-      } else {
-        throw new DdlException(create, "Schema for " + create.name + " not found.");
-      }
-    }
-
-    // TODO: Add support for populating new tables from a query as a one-time operation.
-    if (create.query != null) {
-      throw new DdlException(create, "Populating new tables is not currently supported.");
-    }
-    if (create.columnList == null) {
-      throw new DdlException(create, "No columns provided.");
-    }
-
-    final SchemaPlus schemaPlus = pair.left.plus();
-    String database = null;
-    String tableName;
-    if (isNewSchema) {
-      int idx = pair.right.indexOf(".");
-      database = pair.right.substring(0, idx);
-      tableName = pair.right.substring(idx + 1);
-    } else {
-      tableName = pair.right;
-    }
-
-    if (!isNewSchema && schemaPlus.tables().get(tableName) != null) {
-      if (!create.ifNotExists && !create.getReplace()) {
-        // They did not specify IF NOT EXISTS, so give error.
-        throw new DdlException(create,
-            "Table " + tableName + " already exists. Use CREATE OR REPLACE to update.");
-      }
-    }
-
-    Collection<Deployer> deployers = null;
-    Pair<SchemaPlus, Table> schemaSnapshot = null;
-    try {
-      if (database == null) {
-        if (pair.left.schema instanceof Database) {
-          database = ((Database) pair.left.schema).databaseName();
-        } else {
-          database = connection.getSchema();
-        }
-      }
-
-      final JavaTypeFactory typeFactory = context.getTypeFactory();
-      final ImmutableList.Builder<ColumnDef> columnDefBuilder = ImmutableList.builder();
-      final RelDataTypeFactory.Builder relBuilder = typeFactory.builder();
-      final SqlValidator validator = new ContextSqlValidator(context, true);
-      for (SqlNode columnNode : create.columnList) {
-        if (columnNode instanceof SqlColumnDeclaration) {
-          final SqlColumnDeclaration columnDeclaration = (SqlColumnDeclaration) columnNode;
-          final RelDataType type = columnDeclaration.dataType.deriveType(validator, true);
-          relBuilder.add(columnDeclaration.name.getSimple(), type);
-          columnDefBuilder.add(ColumnDef.of(columnDeclaration.expression, type, columnDeclaration.strategy));
-        } else if (columnNode instanceof SqlKeyConstraint) {
-          // TODO: Support UNIQUE & PRIMARY KEY
-          logger.info("Unsupported column declaration: " + columnNode.getClass());
-        } else {
-          throw new SQLException("Unsupported column declaration: " + columnNode.getClass());
-        }
-      }
-      final RelDataType rowType = relBuilder.build();
-      final List<ColumnDef> columns = columnDefBuilder.build();
-      final InitializerExpressionFactory ief =
-          new NullInitializerExpressionFactory() {
-            @Override public ColumnStrategy generationStrategy(RelOptTable table,
-                int iColumn) {
-              return columns.get(iColumn).strategy;
-            }
-
-            @Override public RexNode newColumnDefaultValue(RelOptTable table,
-                int iColumn, InitializerContext context) {
-              final ColumnDef columnDef = columns.get(iColumn);
-              if (columnDef.expr != null) {
-                final SqlNode validated = context.validateExpression(rowType, columnDef.expr);
-                // The explicit specified type should have the same nullability
-                // with the column expression inferred type
-                return context.convertExpression(validated);
-              }
-              return super.newColumnDefaultValue(table, iColumn, context);
-            }
-          };
-      // Snapshot current state for rollback
-      if (!isNewSchema) {
-        Table currentTable = schemaPlus.tables().get(tableName);
-        schemaSnapshot = Pair.of(schemaPlus, currentTable);
-      }
-
-      // Table does not exist. Create it.
-      // Add a temporary table with the correct row type so deployers can resolve the schema
-      // TODO: This may cause problems if we reuse connections, only the next connection will load this as a HoptimatorJdbcTable.
-      if (isNewSchema) {
-        HoptimatorJdbcCatalogSchema catalogSchema = schemaPlus.unwrap(HoptimatorJdbcCatalogSchema.class);
-        if (catalogSchema == null) {
-          throw new DdlException(create, "Catalog for " + schemaPlus.getName() + " not found.");
-        }
-        SchemaPlus databaseSchema = schemaPlus.add(database, catalogSchema.createSchema(database));
-        logger.info("Added schema {} to catalog {}", database, schemaPlus.getName());
-
-        Table tempTable = new TemporaryTable(rowType, ief, database);
-        databaseSchema.add(tableName, tempTable);
-        logger.info("Added table {} to schema {}", tableName, databaseSchema.getName());
-      } else {
-        Table tempTable = new TemporaryTable(rowType, ief, database);
-        schemaPlus.add(tableName, tempTable);
-        logger.info("Added table {} to schema {}", tableName, schemaPlus.getName());
-      }
-
-      final List<String> schemaPath = pair.left.path(null);
-      List<String> tablePath = new ArrayList<>(schemaPath);
-      if (isNewSchema) {
-        tablePath.add(database);
-      }
-      tablePath.add(tableName);
-
-      Map<String, String> tableOptions = HoptimatorDdlUtils.options(create.options);
-
-      Source source = new Source(database, tablePath, tableOptions);
-      logger.info("Validating new table {}", source);
-      ValidationService.validateOrThrow(source);
-      deployers = DeploymentService.deployers(source, connection);
-      logger.info("Validating deployable resources for table {}", tableName);
-      ValidationService.validateOrThrow(deployers);
-      if (create.getReplace()) {
-        logger.info("Deploying update table {}", source);
-        DeploymentService.update(deployers);
-      } else {
-        logger.info("Deploying create table {}", source);
-        DeploymentService.create(deployers);
-      }
-      logger.info("Deployed table {}", source);
+      HoptimatorDdlUtils.processCreateTable(context, connection, create, mode);
     } catch (SQLException | RuntimeException e) {
-      logger.info("Failed to deploy table {}", tableName);
-      if (deployers != null) {
-        DeploymentService.restore(deployers);
-        logger.info("Restored deployable resources for table {}", tableName);
-      }
-      if (schemaSnapshot != null) {
-        if (schemaSnapshot.right == null) {
-          schemaSnapshot.left.removeTable(tableName);
-          logger.info("Removed schema for table {}", tableName);
-        } else {
-          schemaPlus.add(tableName, schemaSnapshot.right);
-          logger.info("Restored schema for table {}", tableName);
-        }
-      } else {
-        pair.left.removeSubSchema(database);
-        logger.info("Removed schema {} from catalog", database);
-      }
+      logger.info("Failed to deploy table {}", create.name);
       throw new DdlException(create, e.getMessage(), e);
     }
-    logger.info("CREATE TABLE {} completed", tableName);
+    logger.info("CREATE TABLE {} completed", create.name);
   }
 
   /** Executes a {@code PAUSE TRIGGER} command. */
@@ -738,51 +465,4 @@ public final class HoptimatorDdlExecutor extends ServerDdlExecutor {
     }
   }
 
-  private final static class ColumnDef {
-    final @Nullable SqlNode expr;
-    final RelDataType type;
-    final ColumnStrategy strategy;
-
-    private ColumnDef(@Nullable SqlNode expr, RelDataType type,
-        ColumnStrategy strategy) {
-      this.expr = expr;
-      this.type = type;
-      this.strategy = requireNonNull(strategy, "strategy");
-      checkArgument(
-          strategy == ColumnStrategy.NULLABLE
-              || strategy == ColumnStrategy.NOT_NULLABLE
-              || expr != null);
-    }
-
-    static ColumnDef of(@Nullable SqlNode expr, RelDataType type,
-        ColumnStrategy strategy) {
-      return new ColumnDef(expr, type, strategy);
-    }
-  }
-
-  /**
-   * Temporary table implementation used during CREATE TABLE to provide row type information
-   * to deployers before the actual table exists in the underlying database.
-   */
-  private static class TemporaryTable extends ArrayTable<Object[]> {
-    private final InitializerExpressionFactory initializerExpressionFactory;
-    private final String databaseName;
-
-    TemporaryTable(RelDataType rowType, InitializerExpressionFactory initializerExpressionFactory, String databaseName) {
-      super(Object[].class, rowType);
-      this.initializerExpressionFactory = initializerExpressionFactory;
-      this.databaseName = databaseName;
-    }
-
-    @Override public <C extends Object> @Nullable C unwrap(Class<C> aClass) {
-      if (aClass.isInstance(initializerExpressionFactory)) {
-        return aClass.cast(initializerExpressionFactory);
-      }
-      return super.unwrap(aClass);
-    }
-
-    String databaseName() {
-      return databaseName;
-    }
-  }
 }
